requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
git+https://github.com/City-of-Helsinki/django-helusers.git@v0.2.3#egg=django-helusers==0.2.3
cffi==1.9.1               # via cryptography
click==6.6                # via pip-tools
contextlib2==0.5.4        # via raven
cryptography==1.6
defusedxml==0.4.1         # via python3-openid
django-allauth==0.29.0
django-bootstrap3==7.1.0
django-braces==1.10.0     # via django-oauth-toolkit
django-cors-headers==1.3.1
django-multiselectfield==0.1.4
django-oauth-toolkit==0.10.0
django-oidc-provider==0.4.4
django-parler==1.7
django==1.10.3
djangorestframework==3.5.3
first==2.0.1              # via pip-tools
future==0.16.0            # via pyjwkest
idna==2.1                 # via cryptography
oauthlib==1.0.3           # via django-oauth-toolkit, requests-oauthlib
pip-tools==1.8.0
psycopg2==2.6.2
pyasn1==0.1.9             # via cryptography
pycparser==2.17           # via cffi
pycryptodomex==3.4.3      # via pyjwkest
pyjwkest==1.3.0           # via django-oidc-provider
pyjwt==1.4.2
python3-openid==3.0.10    # via django-allauth
raven==5.32.0
requests-oauthlib==0.7.0  # via django-allauth
requests==2.12.1          # via django-allauth, pyjwkest, requests-oauthlib
six==1.10.0               # via cryptography, django-oauth-toolkit, pip-tools, pyjwkest

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via cryptography
